Added possiblity to add new properties

+++ booking-app-backend/src/main/java/aidian3k/project/bookingappbackend/repository/PhotoRepository.java
@@ -8,5 +8,5 @@ import java.util.Optional;
 
 @Repository
 public interface PhotoRepository extends JpaRepository<Photo, Long> {
-    Optional<Photo> getPhotoByPhotoId(Integer photoId);
+    Optional<Photo> getPhotoByPhotoId(Long photoId);
 }
